Take active questions into account when calculating progress for a category.
This change has major impact on the total time required to calculate progress for a category because it requires evaluating pgh_api_question_is_active() for every question in the category. This makes it impractical to do on the fly calculation for the dashboard page. Instead, we now recalculate one category whenever the category form is submitted. We store this value in a serialized array on the Application node. This can be done in a reasonable time and it allows fetching progress to be a simple fast lookup.